+++ index.html
@@ -230,5 +230,6 @@
     </div>
   </template>
 
+<script type="module" src="js/main.js"></script> 
 </body>
 </html>
